Service Bindings find
The Service Bindings find api returns a single binding by name, or
undefined if a binding of the given name does not exist.  Note that this
is different than the filter api which returns an array, which is
potentially empty but if populated, must be used in an indexed fashion.

Signed-off-by: Ben Hale <nebhale@nebhale.com>